Update migration file

+++ database/migrations/2024_10_28_092803_drop_columns_from_settings_table.php
@@ -12,9 +12,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('settings', function (Blueprint $table) {
-            $table->dropColumn(['section', 'subsection', 'hint', 'setting']);
-            $table->primary(['name']);
-        });
+            if (Schema::hasColumns('settings', ['section', 'subsection', 'hint', 'setting'])) {
+                $table->dropColumn(['section', 'subsection', 'hint', 'setting']);
+                $table->primary(['name']);
+}       });
     }
 
     /**
